torchtitan/models/flux2/DataConsolidation/src/avl/requirements.txt
h5py==3.11.0
numpy==1.26.4
PyYAML==6.0.1
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0